+++ src/avdecc/controllerManager.cpp
@@ -18,10 +18,12 @@
 */
 
 #include "controllerManager.hpp"
-#include <atomic>
-#include <la/avdecc/logger.hpp>
 #include "avdecc/helper.hpp"
 #include "settingsManager/settings.hpp"
+
+#include <la/avdecc/logger.hpp>
+
+#include <atomic>
 
 #if __cpp_lib_experimental_atomic_smart_pointers
 #	define HAVE_ATOMIC_SMART_POINTERS
@@ -47,18 +49,30 @@ public:
 
 		protected:
 			// la::avdecc::controller::model::EntityModelVisitor overrides
+			virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::EntityNode const& node) noexcept override
+			{
+				// Initialize internal counter value, always setting lastClearCount to 0 (Statistics counters always start at 0 in the Controller, contrary to endpoint Counters) so that we directly see any error during enumeration
+				{
+					auto const counter = entity->getAecpRetryCounter();
+					_errorCounterTracker._statisticsCounters[StatisticsErrorCounterFlag::AecpRetries] = StatisticsCounterInfo{ counter, 0u };
+				}
+				{
+					auto const counter = entity->getAecpTimeoutCounter();
+					_errorCounterTracker._statisticsCounters[StatisticsErrorCounterFlag::AecpTimeouts] = StatisticsCounterInfo{ counter, 0u };
+				}
+				{
+					auto const counter = entity->getAecpUnexpectedResponseCounter();
+					_errorCounterTracker._statisticsCounters[StatisticsErrorCounterFlag::AecpUnexpectedResponses] = StatisticsCounterInfo{ counter, 0u };
+				}
+			}
 			virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::StreamInputNode const& node) noexcept override
 			{
 				if (node.dynamicModel->counters)
 				{
-					for (auto const& counterKV : *node.dynamicModel->counters)
+					for (auto const [flag, counter] : *node.dynamicModel->counters)
 					{
-						auto const& flag = counterKV.first;
-						auto const& counter = counterKV.second;
-
 						// Initialize internal counter value
-						auto& errorCounter = _errorCounterTracker._streamInputCounter[node.descriptorIndex];
-						errorCounter.counters[flag] = counter;
+						_errorCounterTracker._streamInputCounters[node.descriptorIndex][flag] = ErrorCounterInfo{ counter, counter };
 					}
 				}
 			}
@@ -78,16 +92,30 @@ public:
 
 		protected:
 			// la::avdecc::controller::model::EntityModelVisitor overrides
+			virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::EntityNode const& node) noexcept override
+			{
+				{
+					auto& counterInfo = _errorCounterTracker._statisticsCounters[StatisticsErrorCounterFlag::AecpRetries];
+					counterInfo.lastClearCount = counterInfo.currentCount;
+				}
+				{
+					auto& counterInfo = _errorCounterTracker._statisticsCounters[StatisticsErrorCounterFlag::AecpTimeouts];
+					counterInfo.lastClearCount = counterInfo.currentCount;
+				}
+				{
+					auto& counterInfo = _errorCounterTracker._statisticsCounters[StatisticsErrorCounterFlag::AecpUnexpectedResponses];
+					counterInfo.lastClearCount = counterInfo.currentCount;
+				}
+				emit _manager.statisticsErrorCounterChanged(entity->getEntity().getEntityID(), {});
+			}
 			virtual void visit(la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::controller::model::ConfigurationNode const* const parent, la::avdecc::controller::model::StreamInputNode const& node) noexcept override
 			{
-				auto& errorCounter = _errorCounterTracker._streamInputCounter[node.descriptorIndex];
-
-				if (!errorCounter.flags.empty())
+				for (auto& counterInfoKV : _errorCounterTracker._streamInputCounters[node.descriptorIndex])
 				{
-					errorCounter.flags.clear();
-
-					/*emit*/ _manager.streamInputErrorCounterChanged(entity->getEntity().getEntityID(), node.descriptorIndex, errorCounter.flags);
+					auto& counterInfo = counterInfoKV.second;
+					counterInfo.lastClearCount = counterInfo.currentCount;
 				}
+				emit _manager.streamInputErrorCounterChanged(entity->getEntity().getEntityID(), node.descriptorIndex, {});
 			}
 
 		private:
@@ -111,39 +139,51 @@ public:
 			}
 		}
 
-		la::avdecc::entity::StreamInputCounterValidFlags getStreamInputCounterValidFlags(la::avdecc::entity::model::StreamIndex const streamIndex) const
+		/* ************************************************************ */
+		/* StreamInput Error Counters                                   */
+		/* ************************************************************ */
+		StreamInputErrorCounters getStreamInputErrorCounters(la::avdecc::entity::model::StreamIndex const streamIndex) const
 		{
-			auto const streamIt = _streamInputCounter.find(streamIndex);
-			if (streamIt != std::end(_streamInputCounter))
+			auto counters = StreamInputErrorCounters{};
+
+			auto const streamIt = _streamInputCounters.find(streamIndex);
+			if (streamIt != std::end(_streamInputCounters))
 			{
-				auto const& errorCounter{ streamIt->second };
-				return errorCounter.flags;
+				for (auto const& [flag, errorCounter] : streamIt->second)
+				{
+					if (errorCounter.currentCount != errorCounter.lastClearCount)
+					{
+						counters[flag] = errorCounter.currentCount - errorCounter.lastClearCount;
+					}
+				}
 			}
-			return {};
+
+			return counters;
 		}
 
 		// Set the new counter value, returns true if the counter has changed, false otherwise
 		bool setStreamInputCounter(la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::StreamInputCounterValidFlag const flag, la::avdecc::entity::model::DescriptorCounter const counter)
 		{
-			auto& errorCounter = _streamInputCounter[streamIndex];
+			// Get or create ErrorCounterInfo
+			auto& errorCounter = _streamInputCounters[streamIndex][flag];
 
-			auto shouldNotify{ false };
+			auto shouldNotify = false;
 
 			// Detect counter reset (or wrap)
-			if (counter < errorCounter.counters[flag])
+			if (counter < errorCounter.currentCount)
 			{
-				errorCounter.counters[flag] = 0;
-				errorCounter.flags.reset(flag);
+				errorCounter.lastClearCount = 0u; // Reset counter error (we accept to loose the error state if it was in error, in the case of wrapping)
 				shouldNotify = true;
 			}
 
-			if (counter > errorCounter.counters[flag])
+			// Detect counter increment
+			if (counter > errorCounter.currentCount)
 			{
-				errorCounter.flags.set(flag);
 				shouldNotify = true;
 			}
 
-			errorCounter.counters[flag] = counter;
+			// Always update counter value
+			errorCounter.currentCount = counter;
 
 			return shouldNotify;
 		}
@@ -151,11 +191,12 @@ public:
 		// Clear the error for a given flag, returns true if the flag has changed, false otherwise
 		bool clearStreamInputCounter(la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::StreamInputCounterValidFlag const flag)
 		{
-			auto& errorCounter = _streamInputCounter[streamIndex];
+			AVDECC_ASSERT(_streamInputCounters[streamIndex].count(flag) != 0, "Should not be possible to clear an error flag that does not exist");
+			auto& errorCounter = _streamInputCounters[streamIndex][flag];
 
-			if (errorCounter.flags.test(flag))
+			if (errorCounter.lastClearCount != errorCounter.currentCount)
 			{
-				errorCounter.flags.reset(flag);
+				errorCounter.lastClearCount = errorCounter.currentCount;
 				return true;
 			}
 
@@ -173,16 +214,92 @@ public:
 			}
 		}
 
-	private:
-		la::avdecc::UniqueIdentifier _entityID{ la::avdecc::UniqueIdentifier::getNullUniqueIdentifier() };
-
-		struct ErrorCounter
+		/* ************************************************************ */
+		/* Statistics Error Counters                                    */
+		/* ************************************************************ */
+		StatisticsErrorCounters getStatisticsErrorCounters() const
 		{
-			la::avdecc::entity::StreamInputCounterValidFlags flags; // per flag state 1: error, 0: clear
-			std::unordered_map<la::avdecc::entity::StreamInputCounterValidFlag, la::avdecc::entity::model::DescriptorCounter> counters; // per flag counter value
+			auto counters = StatisticsErrorCounters{};
+
+			for (auto const& [flag, errorCounter] : _statisticsCounters)
+			{
+				if (errorCounter.currentCount != errorCounter.lastClearCount)
+				{
+					counters[flag] = errorCounter.currentCount - errorCounter.lastClearCount;
+				}
+			}
+
+			return counters;
+		}
+
+		// Set the new counter value, returns true if the counter has changed, false otherwise
+		bool setStatisticsCounter(StatisticsErrorCounterFlag const flag, std::uint64_t const counter)
+		{
+			// Get or create StatisticsCounterInfo
+			auto& errorCounter = _statisticsCounters[flag];
+
+			auto shouldNotify = false;
+
+			// Detect counter reset (or wrap)
+			if (counter < errorCounter.currentCount)
+			{
+				errorCounter.lastClearCount = 0u; // Reset counter error (we accept to loose the error state if it was in error, in the case of wrapping)
+				shouldNotify = true;
+			}
+
+			// Detect counter increment
+			if (counter > errorCounter.currentCount)
+			{
+				shouldNotify = true;
+			}
+
+			// Always update counter value
+			errorCounter.currentCount = counter;
+
+			return shouldNotify;
+		}
+
+		// Clear the error for a given flag, returns true if the flag has changed, false otherwise
+		bool clearStatisticsCounter(StatisticsErrorCounterFlag const flag)
+		{
+			AVDECC_ASSERT(_statisticsCounters.count(flag) != 0, "Should not be possible to clear an error flag that does not exist");
+			auto& errorCounter = _statisticsCounters[flag];
+
+			if (errorCounter.lastClearCount != errorCounter.currentCount)
+			{
+				errorCounter.lastClearCount = errorCounter.currentCount;
+				return true;
+			}
+
+			return false;
+		}
+
+		// Clear all the error flags
+		void clearAllStatisticsCounters()
+		{
+			auto& manager = ControllerManager::getInstance();
+			ClearCounterVisitor visitor{ manager, *this };
+			if (auto entity = manager.getControlledEntity(_entityID))
+			{
+				entity->accept(&visitor);
+			}
+		}
+
+	private:
+		struct ErrorCounterInfo
+		{
+			la::avdecc::entity::model::DescriptorCounter currentCount{ 0u }; // Current Counter Value
+			la::avdecc::entity::model::DescriptorCounter lastClearCount{ 0u }; // Value when last Cleared
+		};
+		struct StatisticsCounterInfo
+		{
+			std::uint64_t currentCount{ 0u }; // Current Counter Value
+			std::uint64_t lastClearCount{ 0u }; // Value when last Cleared
 		};
 
-		std::unordered_map<la::avdecc::entity::model::StreamIndex, ErrorCounter> _streamInputCounter;
+		la::avdecc::UniqueIdentifier _entityID{ la::avdecc::UniqueIdentifier::getNullUniqueIdentifier() };
+		std::unordered_map<la::avdecc::entity::model::StreamIndex, std::unordered_map<la::avdecc::entity::StreamInputCounterValidFlag, ErrorCounterInfo>> _streamInputCounters{};
+		std::unordered_map<StatisticsErrorCounterFlag, StatisticsCounterInfo> _statisticsCounters{};
 	};
 
 	ControllerManagerImpl() noexcept
@@ -193,6 +310,8 @@ public:
 		qRegisterMetaType<std::chrono::milliseconds>("std::chrono::milliseconds");
 		qRegisterMetaType<AecpCommandType>("avdecc::ControllerManager::AecpCommandType");
 		qRegisterMetaType<AcmpCommandType>("avdecc::ControllerManager::AcmpCommandType");
+		qRegisterMetaType<StreamInputErrorCounters>("avdecc::ControllerManager::StreamInputErrorCounters");
+		qRegisterMetaType<StatisticsErrorCounters>("avdecc::ControllerManager::StatisticsErrorCounters");
 		qRegisterMetaType<la::avdecc::UniqueIdentifier>("la::avdecc::UniqueIdentifier");
 		qRegisterMetaType<la::avdecc::entity::ControllerEntity::AemCommandStatus>("la::avdecc::entity::ControllerEntity::AemCommandStatus");
 		qRegisterMetaType<la::avdecc::entity::ControllerEntity::ControlStatus>("la::avdecc::entity::ControllerEntity::ControlStatus");
@@ -465,36 +584,31 @@ private:
 	}
 	virtual void onStreamInputCountersChanged(la::avdecc::controller::Controller const* const /*controller*/, la::avdecc::controller::ControlledEntity const* const entity, la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::model::StreamInputCounters const& counters) noexcept override
 	{
-		auto const entityID{ entity->getEntity().getEntityID() };
+		auto const entityID = entity->getEntity().getEntityID();
 
 		if (auto* errorCounterFlags = entityErrorCounterTracker(entityID))
 		{
-			auto const prevFlags = errorCounterFlags->getStreamInputCounterValidFlags(streamIndex);
+			auto changed = false;
 
-			for (auto const& counterKV : counters)
+			for (auto const [flag, counter] : counters)
 			{
-				auto const& flag = counterKV.first;
-				auto const& counter = counterKV.second;
-
 				switch (flag)
 				{
-					//case la::avdecc::entity::StreamInputCounterValidFlag::MediaUnlocked:
 					case la::avdecc::entity::StreamInputCounterValidFlag::StreamInterrupted:
 					case la::avdecc::entity::StreamInputCounterValidFlag::SeqNumMismatch:
 					case la::avdecc::entity::StreamInputCounterValidFlag::LateTimestamp:
 					case la::avdecc::entity::StreamInputCounterValidFlag::EarlyTimestamp:
 					case la::avdecc::entity::StreamInputCounterValidFlag::UnsupportedFormat:
-						errorCounterFlags->setStreamInputCounter(streamIndex, flag, counter);
+						changed |= errorCounterFlags->setStreamInputCounter(streamIndex, flag, counter);
 						break;
 					default:
 						break;
 				}
 			}
 
-			auto const newFlags = errorCounterFlags->getStreamInputCounterValidFlags(streamIndex);
-			if (newFlags != prevFlags)
+			if (changed)
 			{
-				emit streamInputErrorCounterChanged(entityID, streamIndex, newFlags);
+				emit streamInputErrorCounterChanged(entityID, streamIndex, errorCounterFlags->getStreamInputErrorCounters(streamIndex));
 			}
 		}
 
@@ -527,14 +641,44 @@ private:
 	// Statistics
 	virtual void onAecpRetryCounterChanged(la::avdecc::controller::Controller const* const /*controller*/, la::avdecc::controller::ControlledEntity const* const entity, std::uint64_t const value) noexcept override
 	{
-		emit aecpRetryCounterChanged(entity->getEntity().getEntityID(), value);
+		auto const entityID = entity->getEntity().getEntityID();
+
+		if (auto* errorCounterFlags = entityErrorCounterTracker(entityID))
+		{
+			if (errorCounterFlags->setStatisticsCounter(StatisticsErrorCounterFlag::AecpRetries, value))
+			{
+				emit statisticsErrorCounterChanged(entityID, errorCounterFlags->getStatisticsErrorCounters());
+			}
+		}
+
+		emit aecpRetryCounterChanged(entityID, value);
 	}
 	virtual void onAecpTimeoutCounterChanged(la::avdecc::controller::Controller const* const /*controller*/, la::avdecc::controller::ControlledEntity const* const entity, std::uint64_t const value) noexcept override
 	{
+		auto const entityID = entity->getEntity().getEntityID();
+
+		if (auto* errorCounterFlags = entityErrorCounterTracker(entityID))
+		{
+			if (errorCounterFlags->setStatisticsCounter(StatisticsErrorCounterFlag::AecpTimeouts, value))
+			{
+				emit statisticsErrorCounterChanged(entityID, errorCounterFlags->getStatisticsErrorCounters());
+			}
+		}
+
 		emit aecpTimeoutCounterChanged(entity->getEntity().getEntityID(), value);
 	}
 	virtual void onAecpUnexpectedResponseCounterChanged(la::avdecc::controller::Controller const* const /*controller*/, la::avdecc::controller::ControlledEntity const* const entity, std::uint64_t const value) noexcept override
 	{
+		auto const entityID = entity->getEntity().getEntityID();
+
+		if (auto* errorCounterFlags = entityErrorCounterTracker(entityID))
+		{
+			if (errorCounterFlags->setStatisticsCounter(StatisticsErrorCounterFlag::AecpUnexpectedResponses, value))
+			{
+				emit statisticsErrorCounterChanged(entityID, errorCounterFlags->getStatisticsErrorCounters());
+			}
+		}
+
 		emit aecpUnexpectedResponseCounterChanged(entity->getEntity().getEntityID(), value);
 	}
 	virtual void onAecpResponseAverageTimeChanged(la::avdecc::controller::Controller const* const /*controller*/, la::avdecc::controller::ControlledEntity const* const entity, std::chrono::milliseconds const& value) noexcept override
@@ -660,11 +804,11 @@ private:
 		return const_cast<ErrorCounterTracker*>(static_cast<ControllerManagerImpl const*>(this)->entityErrorCounterTracker(entityID));
 	}
 
-	virtual la::avdecc::entity::StreamInputCounterValidFlags getStreamInputErrorCounterFlags(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::StreamIndex const streamIndex) const noexcept override
+	virtual StreamInputErrorCounters getStreamInputErrorCounters(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::StreamIndex const streamIndex) const noexcept override
 	{
 		if (auto* errorCounterTracker = entityErrorCounterTracker(entityID))
 		{
-			return errorCounterTracker->getStreamInputCounterValidFlags(streamIndex);
+			return errorCounterTracker->getStreamInputErrorCounters(streamIndex);
 		}
 		return {};
 	}
@@ -675,8 +819,7 @@ private:
 		{
 			if (errorCounterTracker->clearStreamInputCounter(streamIndex, flag))
 			{
-				auto const flags = errorCounterTracker->getStreamInputCounterValidFlags(streamIndex);
-				emit streamInputErrorCounterChanged(entityID, streamIndex, flags);
+				emit streamInputErrorCounterChanged(entityID, streamIndex, errorCounterTracker->getStreamInputErrorCounters(streamIndex));
 			}
 		}
 	}
@@ -686,6 +829,34 @@ private:
 		if (auto* errorCounterTracker = entityErrorCounterTracker(entityID))
 		{
 			errorCounterTracker->clearAllStreamInputCounters();
+		}
+	}
+
+	virtual StatisticsErrorCounters getStatisticsCounters(la::avdecc::UniqueIdentifier const entityID) const noexcept override
+	{
+		if (auto* errorCounterTracker = entityErrorCounterTracker(entityID))
+		{
+			return errorCounterTracker->getStatisticsErrorCounters();
+		}
+		return {};
+	}
+
+	virtual void clearStatisticsCounterValidFlags(la::avdecc::UniqueIdentifier const entityID, StatisticsErrorCounterFlag const flag) noexcept override
+	{
+		if (auto* errorCounterTracker = entityErrorCounterTracker(entityID))
+		{
+			if (errorCounterTracker->clearStatisticsCounter(flag))
+			{
+				emit statisticsErrorCounterChanged(entityID, errorCounterTracker->getStatisticsErrorCounters());
+			}
+		}
+	}
+
+	virtual void clearAllStatisticsCounterValidFlags(la::avdecc::UniqueIdentifier const entityID) noexcept override
+	{
+		if (auto* errorCounterTracker = entityErrorCounterTracker(entityID))
+		{
+			errorCounterTracker->clearAllStatisticsCounters();
 		}
 	}