+++ src/counters/streamInputCountersTreeWidgetItem.hpp
@@ -36,7 +36,7 @@
 class StreamInputCounterTreeWidgetItem : public QTreeWidgetItem
 {
 public:
-	StreamInputCounterTreeWidgetItem(la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::StreamInputCounterValidFlag flag, QTreeWidgetItem* parent)
+	StreamInputCounterTreeWidgetItem(la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::StreamInputCounterValidFlag const flag, QTreeWidgetItem* parent)
 		: QTreeWidgetItem{ parent }
 		, _streamIndex{ streamIndex }
 		, _counterValidFlag{ flag }
@@ -63,8 +63,6 @@ class StreamInputCountersTreeWidgetItem : public QObject, public QTreeWidgetItem
 public:
 	StreamInputCountersTreeWidgetItem(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::StreamIndex const streamIndex, la::avdecc::entity::model::StreamInputCounters const& counters, QTreeWidget* parent = nullptr);
 
-	void setStreamInputErrorCounterFlags(la::avdecc::entity::StreamInputCounterValidFlags const& flags);
-
 private:
 	void updateCounters(la::avdecc::entity::model::StreamInputCounters const& counters);
 
@@ -72,5 +70,7 @@ private:
 	la::avdecc::entity::model::StreamIndex const _streamIndex{ 0u };
 
 	// Counters
-	std::map<la::avdecc::entity::StreamInputCounterValidFlag, StreamInputCounterTreeWidgetItem*> _counters{};
+	std::map<la::avdecc::entity::StreamInputCounterValidFlag, StreamInputCounterTreeWidgetItem*> _counterWidgets{};
+	la::avdecc::entity::model::StreamInputCounters _counters{};
+	avdecc::ControllerManager::StreamInputErrorCounters _errorCounters{};
 };